Add temp termite transparency, more useful binds
The current solution to make Termite transparent is to draw its entire
window class with set transparency (in picom config). This decreases the
transparency of the text too, which is unwanted. For some reason
changing the background transparency inside Termite config didn't work,
I will look into this issue in more detail later.