Tweak terminal colors smore, fix i3 titles
I've replaced terminal pink with orange and made the blue closer to the
sea color (between blue and green).
The i3 fixes are very minor. In the last snapshot I made scratch term
have a custom title displayed, but I've forgotten to also add it to the
auto-spawned scratch term at startup. This has been fixed of course.