Patch a minor glitch in i3
Basically, while doing the resizing or movement with scripts, it's
possible to glitch it out and cause the window to infinitely grow or
shrink. Patched by introducing more ubiquotous --release conditions for
killall commands.
I haven't managed to find a way to patch the same glitch from occurring
with movement, but it's not a big deal, I'll let it slide
(literally, haHAA).