Change thunar bind to lf, update desktop app names
The "generic" desktop files (file, text, image) now in their Name field
store the actual name of the program that's being used. This is an
aesthetic change so that various parsers that poll information about
which program opens what actually display the name of the program, as
opposed to something generic like "Text Editor".

Ctrl+Alt+I now opens a new LF session instead of Thunar.