Minor tweaks to crontab and sysupdate
sysupdate will now run yay with -a flag, causing it to only care about
AUR.

I found out crontabs aren't as clever as I thought they were, and it's
impossible to easily run a job every 45 minutes, so I changed exercise
prompts to be run hourly.

Put a newline between program options and LF icons in profile, because
it was inconvenient to navigate through in vim.