Remove junk languages from beets preferences
Well, not exactly "junk", since I deliberately put them there, but it
appears that beets does not distinguish well the hierarchy, or order in
which these languages are picked, i.e. I was getting japanese titles
everywhere until I removed the "ja" preference. So, the other ones went
out the window as well. Maybe it's worth creating an issue on github
sometime, but for now I didn't feel like it.