Introduce delay when opening lf in dualf panes
For some reason lf sometimes doesn't register the variables set in the
shell when starting dualf. This issue goes away when exiting and
reopening lf in a pane, which leads me to believe that it's just trying
to be opened too fast. So, as a fix, I added a small delay between
setting the variables and opening lf, and so far it seems to have fixed
the issue.