Reformat lfrc scripts, change dual match to dupe
I've done the same kind of reformatting that I was doing on my
standalone scripts very recently.

The former dual-match function used to change the current panel to
whatever the other panel was displaying (lf running in dual mode).
I found that largely inconvenient and counter-intuitive after a few
uses, so I've rewritten the function to dual-dupe, which does the
opposite: changes the other panel to the current selection. It's both
more useful, and significantly simpler in execution, so that's good.