Add mp3 information to fls
When I first designed fls, it was meant to just display basic file
information. This is still the case, except in this snapshot I've added
"extras", which is filetype-specific information.
I will surely add more support for different filetypes in the future,
for now I've just added it for "audio/mpeg" MIME type (mp3).