Add odpush and ditch MEGA cloud for OneDrive
My former setup was using MEGAsync and simply rsyncing files into my
~/MEGA directory and letting the background process handle the uploading
and all. While this worked, MEGAsync is a terrible piece of software can
take literally hours to compile from source code (C++ and only from
AUR). This is an absolute nightmare on lower-end machines and I don't
require most of the robustness the software offers anyway.

I'll be on the lookout for my own cloud solution, with a VPS or
something alike, but for now I chose to exploit OneDrive, because
I happen to have a whole 1TB of storage available there, and since it's
one of the bigger players in the cloud industry, it has relatively good
support of various open-source projects.

I've written a script with one of those projects, and that script
(odpush) basically allows me to list any files/directories I want, and
for each of those paths it will
  1) rsync the contents to a ~/OneDrive/Sync local directory
  2) push the contents of that directory to my remote OneDrive cloud
Turns out, this setup works really well and fulfills my every
expectation for a one-way synchronization with a remote server.
Downside? There isn't a simple way of downloading the remote files, for
now I'll stick to just pulling them from the website GUI when I really
need to.