Add support for wacom device, integrate gromit-mpx
gromit-mpx does not allow for a nice integration with a drawing tablet,
unfortunately. Most shortcuts have to be used in conjuction with a
physical keyboard as opposed to the built-in buttons on the
tablet/stylus. This is because gromit-mpx only allows reading from
buttons 1-5, and the wacom device is best configured to use buttons from
outside the existing ranges (11+) so that no shortcuts collide. Even
when I tried setting the wacom keys to use values within the 1-5 range,
it didn't work as expected and was very cumbersome. Pity.

Added template scripts for bindings wacom tablet functions. These will
later be filled with whatever I need.