Add toggling RMB drawing to wacom config
I originally wanted the functionality of being able to hold one of the
buttons on the pen to enable right-click drawing, and the plan to do
that was to set up xbindkeys to run one script on button press, and
another on button release. I managed to set it up and it was running
both scripts correctly, however I wasn't able to draw and hold the
button simultaneously, must be a driver limitation or something else I
can't really fix.

To counter this, I simply made the button toggle "RMB mode" on and off
with each click, which might prove even more convenient, given that it's
cumbersom to have to keep holding the button while drawing.