Replace polybar newsboat with i3 scratch window
Polybar no longer displays the number of unread articles.
Added a "dualboat" script for spawning a tmux session with both newsboat
and podboat.
Added i3 shortcuts for spawning and toggling a scratch dualboat window.
Removed the nb alias for newsboat, because instead of typing I'll be
using the scratch window.