Tweak newsboat and podboat's keyboard shortcuts
Some shortcuts were annoying to use, so I simplified them.
The biggest change is remapping P for pb-purge to r, because podboat
calls this operation "Purge finished", but in actuality what it does is
just reload the queue file, so I bound it to r for "refresh" or
"reload". Seems more intuitive.