Reconfigure newsboat and integrate with polybar
The previous attempt to integrate newsboat with polybar relied on
newsboat's "-x print-unread" option, which output the number of unread
articles. The problem is, it only works if there are no existing
instances of newsboat running. That was annoying. because the polybar
number would disappear when opening newsboat and stuff like that.

This commit's solution to that is way better. Newboat has a built-in
option to supply a notification script every time the feeds are updated.
What I did was write a script that parses and drops the number of unread
feeds and articles into a polybar-formatted text file, and then polybar
reads that file and diplays it. Now the integration is seamless.

Other changes:
- removed the obsolete terminal_run_newsboat script
- reverted some earlier keybinding changes (P for purging in podboat,
because it turns out I can't have "r" assigned to both pb-purge and
reload, unfortunately)
- Removed newsboat updating as a cronjob (redundant, because I always
keep an instance open as a scratch window now)