Add music library normalization script
What would usually happen is, after importing music I would go to
~/Music, and issue the following command manually:

  find . -type f -name '*.mp3' -print0 | xargs -0 mp3gain -s i -r -e -c

This command runs mp3gain on every single mp3 file in my library, which
works beautifully. However, it's pretty long and annoying to type, so
the new script handles that for me with a surplus. It will also generate
a log file containing each change applied to a file, and, if non-empty,
it will automatically open the log with the default pager. Very neat for
getting a conclusive report on the command's doings.