Change podboat player from mpv to mpdplay
mpdplay is a script I wrote for playing arbitrary music files through
MPD by inserting tracks into the queue. Using MPD for podcasts has the
advantage of being able to use system-wide shortcuts for manipulating
the playback. A drawback is not being able to play music at the same
time, but I've never had the need to listen to music and podcasts
simultaneously, so I'm not expecting it to become a problem.