Add desc to normalize script, revert dualboat keys
I've added a new block of comments in the normalize script to explain
the complete steps necessary to normalize the music library's volume
levels based on replaygain tags. It can be tricky, so I figured it's
better to put it there instead of expecting the user to guess how to use
the script.

In a previous commit I've changed the dualboat exit key combo from C-q
to q, but it turns out q doesn't work well, because I need to press q to
for example stop playback of mpv.