Make pladd compatible with ncmpcpp playlist adding
The pladd script would insert a newline, and then a path. Adding a
playlist item through ncmpcpp after pladd would make two paths appear on
the same line, which is an error.
The solution was to insert first the path, then a newline in the pladd
script. Now both methods work together.