Move root scripts to ~/.local/bin for portability
I found that I have no scripts that necessarily need to be global to the
entire system, so it's way better to have them in my home directory.
Suppose I want to install my setup on someone else's computer where I
don't have all the root privileges, it would be inconvenient to
"require" access to the /usr/local/bin directory.