Add mlocate-like scripts for filesystem scouting
I found that the actual mlocate package is needlessly complicated and
uses some mechanisms that I didn't feel like learning at all, so instead
I wrote my own implementation based on find and grep. It's very
flexible and, what's the most important, it's simple and I understand
the way it works.

Added an entry to crontab to update the database daily.