Autostart tmux in tty, move tmux config
Whenever I find myself in the tty, I forget to turn on tmux and then I
have no tabbing/splitting features which are so handy. So tmux now
autostarts in any tty outside of tty1, where xorg is started by default.

tmux is stupid enough not to be compliant with the XDG standard, so the
only reliable way of loading the config file from a different location
is by aliasing the -f option and specifying the config file with it.

Since my previous binding for reloading config from within tmux didn't
seem to work no matter what path I supplied it, I got rid of it
completely. I don't think I've used it even once so it's not a big loss.