Strip home directory of most programs' files
I've gone through this detailed list on the Arch Wiki:

https://wiki.archlinux.org/index.php/XDG_Base_Directory

and managed to cut the number of files and directories in $HOME by more
than half, reducing from around 80 to 32. As far as I'm aware I've
tested everything and it works properly, but follow-ups may happen as I
discover issues or move away more directories.