Revamp fairuse scripts to follow new conventions
Everything about each script is laid out in comment blocks, so there's
no point in repeating it here.

I have one more script to write, `genslic`, which will generate complete
license information with respect to each track and artist.

I've decided that not all metadata will be stored directly inside the
mp3 music files. Links to artists' web pages will be stored separately
in a yaml file alongside playlists. This is better, because it means
less noise in the id3v2 tags, and less redundancy (each artist has a
fixed set of links, so pasting those links in every single song is a bit
of an overkill).