Disable beets genre whitelist
Beets has been particularly bad at tagging music genre recently, and
that's because in the configuration I set it to force music genre to
comply with the built-in whitelist. That is ridiculous, since many
artists on pages like SoundCloud come up with their own genre names that
befit their music better. And the built-in whitelist fails to include
some niche genres.

From now on, beets will only find genres for music that doesn't have any
genre assigned to it, and leave all original tags in tact.