Fix Xresources syntax and pywal script
The correct syntax for suckless terminal is "St", not "st".
The script needed to be fixed because it would change all
output to lowercase, whilst it is only necessary to change
the color hex codes to lowercase (well, not necessary, but it's an
aesthetic choice that I've come to follow in my Xresources file).