Start dwmblocks in the main dwm init script
Previously both dwm and dwmblocks would initialize shared memory, but
now I've changed it so only dwm can, which enforces dwmblocks to be
started up later. This is for the better, it outlines a clear hierarchy
and makes everything more stable, as dwmblocks has to write its PID to
sharedmemory etc.