Silently exit passcpy and usercpy on empty values
When I originally wrote these scripts I chose not to exit regardless of
whether or not the sensitive data is empty, because in some obscure
cases it's possible that the password/username is actually blank, so
it's a potential vulnerability to have the scripts handle these cases
differently. But this problem is outweighed by the fact that if you
exit out of the gpg-agent pinentry prompt e.g. by pressing escape,
you STILL get the notification claiming something has been copied, which
is annoying.