Revert default browser from surf to brave
Surf has some very significant issues on top of being built on webkitgtk

- it messes up master volume every time it attempts playback
- it crashes frequently on heavy sites like YouTube
- it does not have good support for copy+paste, etc.

HOWEVER it is still REALLY good for opening up light websites like
documentations and keeping them as separate windows with
vim-keybindings. For that reason I've decided for now to keep brave as
the default browser, but also have surf as a lighter alternative.