Remove Projects directory from home tree
I used to organize my files like this:

~
    Documents
    Pictures
    ...
    Projects
        Documents
        Pictures
        ...

basically, the ~/Projects directory would contain my OWN work, while the
regular home directories would contain random/temporary files. This
worked great for backups, but is tedious and unintuitive to use, because
there's a lot of additional decisions involved, and similar files are in
multiple locations, etc. That's why I've moved on to another system:

~
    Documents
        0wn
    Pictures
        0wn
    ...

In this scheme, every filetype resides in its ONE respective directory.
Things that I make myself, projects and other things, always reside in
an "0wn" directory (spelled with 0 so that it is displayed on top in
sorted view).