Disable nvim C-f and C-b navigation mappings
C-f is bound to various FZF actions, so having C-f mapped to anything
causes it to run if you press C-f and wait for the delay to timeout.
This is confusing and annoying and I got rid of it. Also removed C-b
because it doesn't make much sense to have it without having C-f.
Also, I pretty much always use C-d and C-u for quick navigation, so I
don't have much use for C-f and C-b anyways.