Add spull script and update spush
spull is complementary to spush, it distributes files from ~/Sync all
over the system. For that reason it has to be executed with care, only
if the user is absolutely sure that the files stored in ~/Sync are
up-to-date.

Added "-e" flag to both scripts, which is a shortcut for opening their
respective source list files in $EDITOR.

Removed "-t" and "-u" flags from rsync in both scripts, because they
don't matter and only add a layer of complexity.