Add newsboat-mpv-browser script for notifications
The script is an intermediary between newsboat and mpv and basically
sends a notification describing the youtube-dl error when something goes
wrong. For example, when trying to play a youtube video from newsboat,
if the video is set to a premiere, a notification will appear instead of
newsboat just silently failing to do the task. This is just a small
quality of life improvement.