[scripts] Add image support to mdtopdf script
I rewrote the script (if it can be called that) from scratch.

wkhtmltopdf now embeds images specified as relative/absolute paths to
the local filesystem, which is nice (especially with vim previews).

The downside is that a temporary file at the location of the original
markdown file has to be created. This is because for some reason
wkhtmltopdf assumes image paths start at /tmp whenever you feed it code
through standard input, so instead I had to feed it an actual file.