Replace slop back with xrectsel
slop has an amazing option of selecting an entire window by clicking,
but the fact that the selection process is interruptable by pressing any
keyboard button disquilifies this program in my eyes, unfortunately.
xrectsel, on the other hand, ignores keyboard presses, which makes it
possible to combine with programs like mousemode.