[scripts] Revamp note script once again
This time the change is more subtle -- from now on all notes' archives'
top directory is always called "NOTE". The former naming convention
worked as follows:

archive: example.txt
archive contents:
    [DIR] example.txt
      `--NOTE.txt

However, this structure relies on the fact that archive's name is
exactly the same as the name of the top directory within that archive.
This makes it impossible to rename the archive without breaking the
integrity of the note, and is a very poor and redundant design choice.

From now on, all notes' structure looks like this:

archive: example.txt
archive contents:
    [DIR] NOTE
      `--NOTE.txt

(NOTE may still have other extensions like *.md, *.ms, etc.).
This simplifies the implementation a bit and makes renaming archives
possible with no repercussions.

Furthermore, it is no longer necessary that archive names have file
extensions, because the note format is inferred from the suffix of the
NOTE.* file within the NOTE directory. This means you can manually
rename archive files to anything you like, BUT the note script is
incapable of creating a new note without one of the extensions. This is
a person preference - I like having my archives suffixed to have an idea
of what sits within.