[nvim] Improve markup screenshotting functions
Now the process is more interactive; the former approach was to fill a
line of a buffer with the filename, and then after pressing the
keybinding vim would run the screenshotting utility through a system()
call and you would have to select a rectangle area to be save as a
screenshot.

The new approach is better: you simply press the screenshot binding, vim
prompts you for a filename and then pulls the image from the clipboard,
along with many helpful error checks along the way. Situations like
    1) no image in clipboard
    2) image already exists. overwrite or reuse?
are handled interactively or automatically depending on the need.
Overall it's a better experience.