[scripts,nvim] Fix bugs related to notes w/ images
The major change is, you can now successfully embed images into your
notes with vim (see sshot.vim). This functionality was available for a
long time, but it didn't work specifically for notes, and I finally
identified the issue (the issue was, the former ntmake script relied on
the fact that the working directory be the directory where the note in
question resides, but it has been changed in this snapshot).

A few miscellaneous changes:
- simplified some functions in sshot.vim (less arguments)
- added -p flag to groff and pdfmom for note compilation
- removed some redundant chmod calls (I didn't realize mktemp sets
permissions on its own)
- ntmake now demands either - or filename as first argument (this allows
it to cd into the proper directory before compiling notes, so the
filename option is preferred).