[shell] Add xpo to aliasrc
xpo is an alias for setsid, to run something in a new window.
The name "xpo" is meant to be like "xo" (xdg-open), but with "p" for
"pop out". It doesn't really make sense logically, since xdg-open and
setsid work very differently, but it sits right inside my head, and
that's what matters.