Move some cronjobs to anacron
Some of the things I wanted from my crontabs did not work well on
workstations that are not booted up most of the time (like laptops
etc.). I was looking for a fix and found out about anacron, which is the
perfect tool for the job.

Some of the jobs that have been moved:
- weekly clearing of pacman/yay cache
- daily cleanup of old neovim backup files
- weekly playlist management scripts (ran in user mode)