[scripts] Small ncmpcpp-song-change.sh fix
This commit is a follow-up to 8c976a54af35bfeb0d98b60b0a23d9493fa74ad6

Basically, ffmpeg used exit with a non-0 code, which made it possible to
put it in an if statement and based on that either update the wallpaper
to the cover image, or bring back the default.

mpc always succeeds and it seems to fill the file with text upon
failure, so to determine whether an image exists or not, instead of
relying on the exit code I simply grep the output of "file" command, and
if it contains "image" inside, I assume it's an image.

Also I added some safety double quotes around XDG_CACHE_HOME expansion
and broke the code into more reasonably-sized lines.