Migrate from PulseAudio to PipeWire
I've been using PipeWire for a couple days and it seems to be working
fine, with the added bonus of patchbays etc. I had to transition
from vanilla PulseAudio to either JACK/PulseAudio or PipeWire, because
I've been wanting to get into music production on Linux. JACK/PulseAudio
setups are more mature, but sketchy and old. Embrace the future that is
PipeWire!