[lf] Remove dualf, add chmod command
I have never, ever used dualf for anything, and frankly I don't trust it
as a reliable file manager, there's probably many bugs in it. So, I
removed it. "Vanilla" lf is enough for me.

I added a chmod command to lf, which allows to modify permission flags
on a single file or a selection of files. Bound to "P".