Start messing around with River WM
I was hoping to migrate to Wayland, and thought River would be a good
candidate for a WM. Unfortunately, there are many features missing and a
lot of support that I can't quite work around yet.

Still, I like where River is going and I am likely going to revisit it
once some more time passes and it matures. In the meantime, all the
configuration that I managed to do so far will be safely kept on this
new branch.

I will probably try sway next, just because it is the biggest tiling WM
for Wayland out there, so I'll get a good understanding of what the
upper and lower bounds of Wayland options are there on Linux currently.
And if it works and I like it enough - who knows, maybe I'll migrate?
Even though I'm not a big fan of i3.