[hlwm] Optimize and simplify scratchpads
I had an epiphany that, instead of spawning a listener to detect when a
scratchpad closes, we can simply use "wait $!" instead, since we are
inside the process which spawns the scratchpad. This is MUCH simpler and
speeds things up (also removes the need for a constant listener for each
scratchpad in the background, which is good for the speed of ALL scripts).

The cost for making this change was removing the "remap" action which I
added yesterday. "remap" is impossible to implement in a simple way
after this change, because it would require killing the waiting process
to cut the connection between an unmapped scratchpad and a corresponding
my_spwid_* attribute.
I'm not so sure "remap" was such a great idea anyway, its practicality
is debatable (and is outweighed by the implementation complexity).