[hlwm] Simplify layout saving and loading
Enumerating layouts in layout_load.sh based on tmod is a very poor
choice, because tmods vary when you move files between different
machines.
To ensure that each layout has a unique number, the simplest solution is
to include that number in the name of the layout itself, hence the
filename changes.

Both layout_load.sh and layout_save.sh now sort layouts based on
filename before showing them in dmenu. This guarantees they always
display the same thing.
It is up to the user to prepend a number or anything else to sort by
during the saving of a layout.