[bspwm] Fix issues with virtual monitors
The polybar workspaces module, scratchpads and other misc things are now
compatible with running multiple monitors (well, I actually haven't
tested it on physicalmultiple monitors, but virtual ones, both
overlapping and non-overlapping, seem to work just fine).

Scratchpads are shared by all monitors, which is an obvious thing to me.
Polybar shows the name of the current monitor and only workspaces
available on the currently selected monitor. This is probably not ideal
for different physical monitors (because your monitors show the same set
of workspaces as opposed to sticking to their own ones), but works great
for several virtual monitors on the same display.