[sxhkd] Enhance desktop cycling keybinds
Before, super + brackets would cycle all local occupied workspaces.
I was annoyed by how the "_sp" workspace was included in that set
(because it's meant to be a hidden workspace).

This commit adds a new script which cleverly checks for any _*
workspaces and skips them, giving the desired perfect behavior.