[scripts] Improve album art wallpaper generation
With help of imagemagick, I extended the script to create a 16:9
wallpaper which in the center puts the album art, and fills the
background with gaussian-blurred version of the same image.

This took some research to figure out, but is surprisingly simple to
achieve, I never really knew how great imagemagick is.