[polybar] Add more events to bspwm_xwindow module
added node_transfer, the module will now update when sending a node out
of the current desktop.

To be honest, the bspwm_xwindow module is not perfect and never will be
unless I rewrite it in C with xcb/Xlib, which I'm sort of wanting to do
because it would make it much more efficient and functionally sane.