[scripts] Revamp plfilter script
The interface is copied from plcategorize. It is now possible to create
a custom selection of green and red playlists, and the final result will
include songs which belong to all the green ones and none of the red
ones.
The script speed has increased dramatically by using lua for hashsets.