[nvim] Add latex graphviz support and more
The graphviz CRAN package was a bit of a pain to figure out, because it
isn't very consistent or flexible with where it expects certain files to
be. However, with some script trickery I managed to make it stable, or
so I hope. This means embedding dot graphs is now easy and does not
require any external files!

Also added other, smaller changes.