[scripts] Add dirty hack for minted in lxmake
The problem is, minted needs the package outputdir option to be equal to
-output-dir. I used to solve it by conventionally using \x13OUTDIR\x13
in my source file and using sed to substitute it to -output-dir (in
lxmake). However, the problem is that now the source file might be a
subfile, with the main file residing somewhere else. In that case, the
main file would also need the substitutions, and currently I added a
dirty hack to do that, but it's unsafe and should be rewritten in the
future.