[scripts] Make dirty hack in lxmake less dirty
I discovered an important thing: subfiles can be nested, but all of them
always point to the same root document, (as opposed to their "branch
parents").
This means that, with some assumptions, a subfile compilation can be
assurred if we create a mock copy of the root document, substitute
\x13OUTDIR\x13 there, relink the subfile to the mock document, compile,
and then clean up.
The new version of the script does exactly that and it seems to be
working correctly.