[nvim] Add latex keymap for new subfile
It cost me a lot of time and patience, but I've written a vimscript
function that basically intakes a path relative to the opened latex
document, inserts an appropriate \subfile{...} line, creates a new
buffer of the subfile, drops a template inside (with automatic routing
back to the main file) and positions the cursor to be ready to edit.
This should make creating multi-file documents much more enjoyable.
Definitely more enjoyable than writing this function was.