[scripts] Rewrite plrare-stats calculations in lua
The biggest improvement is the utilization of a cache file for song
durations - this cuts down runtime IMMENSELY, as you no longer have to
run mp3info on every single analyzed track.

Other than that, luatables made it easy to implement a simple top-10
listened artists and albums, which is a very welcome addition to the
stats.