[scripts] Add top artists/albums time percentage
The script now displays the percentage of total time the top 10 artists
and albums constitute. This gives some extra insight into how big of a
role those top positions played, which is useful when i've listened to a
lot of different stuff instead of focusing on a few big releases.