[lf] Fix(?) fuzzy find with weird characters
fuzzy-find and fuzzy-find-recursive used to not work when files or
directories contain characters like ' or ". I don't know how lf parses
its arguments exactly, I've tried different things and I settled on
escaping every character. Seems extreme, but also it seems to work.

e.g. If the directory name is ' 'a$bc", it will be converted to a

	:select \'\ \'\a\$\b\c\"

command. This depends on the -E flag in sed, which is not
POSIX-compliant, but I currently do not care.