[scripts] Improve counting on ncmpcpp-song-change
The former script simply set off a timer every time a song changed.
This worked fine, but if you were to switch back and forth in the
playlist, you would spawn multiple timers which sometimes lead to
counting more plays than have actually occurred.

The new implementation uses a simple cache file for the last subshell's
PID and on each song change the last subshell is killed (if still
running ofc). This effectively gets rid of the issue and it's safe to
skip around in the playlist as much as you please now.