[ncmpcpp] Fix song change PID mechanism
My old code actually stored the PID of the parent process, not the
subshell, making it COMPLETELY useless. This time I actually got it
right and I also moved the file location to tmpfs for quicker access
(RAM). While not perfect, this ensures sufficient protection against
running the script in quick succession for my use case.