[scripts] Attempt to fix history-playlists-sync conflicts
In particular, my hist.phone.m3u keeps causing sync conflicts. I'm
pretty sure this is because of the update on the modified timestamp
caused by the sed reformatter and by copying. So, I added some timestamp
preserving flags to cp and I only perform the sed substitution if the
file contains stuff that should be removed. I'll see if these changes
fix it or not.