[scripts] Make lxmake always compile 3 times
I discovered that the lastpage package gets the numbering wrong if the
number changes from 1st to 2nd compilation (e.g. as a result of
inserting the table of contents, which can change the entire document's
layout). Compiling 3 times solves this. This is a big additional time
cost, but I'm willing to make it for the sake of consistency.