[scripts] Rework note -t flag to -r/--raw
Since notes are not necessarily single-file anymore (and they haven't
been for a long time), there is no point having an option that outputs
to a single text file. I changed it to -r or --raw, which outputs the
entire note directory. That way this can actually be shared with other
people.

This removes functionality of viewing a note in textonly form. Again, it
doesn't make much sense to do so anymore. Just edit the note if you want
to see the source.