[scripts] Fix note tex cache generation
I overlooked the necessity to perform string substitution on the latex
source before compiling the cached document. While fixing this, I also
added some clarification for using strsub, which was not obvious
beforehand (it relies on git log, so the `source` argument must point to
a file inside the notes git repo).