[scripts] Add copy-links utility script
I had to save a lot of links from a website today, which due to js
couldn't be done by saving the HTML file and grepping through it. So, I
wrote this script which monitors the clipboard and grabs its contents
when it change. This speeds things up significantly and maybe I will use
it again.