[picom] Enable vsync, disable misc other settings
This fixes a major problem I introduced in
787a2bfd226f4eb3beeb8274d7d35193aeb360be by disabling vsync. For some
applications, most notably vim, the screen would not be updated until
some other action forced an update (e.g. a refresh of the status bar,
or focusing a different window). Since I had a second-count on my status
bar, I observed this as split-second stutters of the entire application,
which was infuriating. Luckily I was able to pin it down today.