Migrate from spectrwm to dwm
Yes, I'm coming back to dwm after a month of using spectrwm as my daily
driver. The reason was that I was unable to perfectly reproduce some of
the behaviors that I found important in my workflow.

For example, my scratchpad script with spectrwm works, but whenever a
scratchpad is shown, it gets appended to the end of the window stack and
focused. Then, when I hide it, the last window in the stack claims the
focus. This is aggravating when working on the master window, and using
the scratchpad for quick swap-in action.

There are several other quirks that led to the ultimate downfall of
spectrwm. Throwing and resizing windows using EWMH randomly stops to
work sometimes, which I can't explain in any other way than blaming
spectrwm for not supporting events correctly.

dwm had some problems in the past, but I've done a clean build, much,
MUCH simpler than the first one I had a few years ago. I hope stability
will no longer be an issue this time.