[scripts] Add video format picker to rss-mpv-browser
ytdl by default picks the best video quality available, which is a
problem e.g. when watching a video that has the best quality of 4k on an
old laptop.
This gets rid of the problem by presenting the user with a video format
picker. It is now possible to select 1080p, 720p, or anyp with ease.

A downside is that this requires an extra ytdl command at the start to
fetch the available formats. This means starting up a video now has an
extra overhead of a second or two. Whether this is annoying, I'll have
to see.