[scripts] rss-mpv-browser: Sort + default video quality
I've added an option to simply use the default video encoding. This is
sometimes useful, as yt-dlp is not perfect when it comes to fetching
available ones (I've had it happen that for some sources, other than
YouTube, yt-dlp yields a long list of video options, none of which are
actually supported).

Also, the video options are now sorted desc by approximated filesize.
This is rudimentary, not perfect (because the output of yt-dlp is not
split into perfect columns), but it works well enough to be very useful.
Now the interesting options start at the top, there is rarely a reason
to pick an option that's not within the first 5 entries.