[scripts] Fix note script ledger support
Ledger has a kickass feature that lets it read GPG-encrypted journals.
Unfortunately, my note script stores notes as GPG-encrypted DIRECTORIES
of possibly many files. This is more flexible, but is not understood by
ledger. So, it was a bit of a pain and made the script uglier, but I
wrote in special ledger support that stores ledger notes as encrypted
files.
This means I can keep my finances encrypted along with the rest of my
notes, and simultaneously use ledger without hassle. Epic!