[shell] Tweak jumprc 'le' alias to open ledger annual file
I wanted to go with monthly files, but I think annual are a better idea,
mostly due to the fact that this will result in fewer files. And since
ledger has to GPG-decrypt every file to read the full journal, the
slowdown could become significant after a few years (when there's a few
dozen files).