[scripts] Export TYPST_ROOT when editing a note
This somewhat helps when a note is a multi-file typst document.
Typst does not have a good workflow for subfiles yet, unfortunately.
For example, when you compile a subfile, the subfile will not inherit
styling from the main document. So the output looks all kinds of wrong.
But exporting TYPST_ROOT at least eliminates errors with nested paths,
etc.