[scripts] Remove neatroff from api, obsolete ntmake
I don't and have not used neatroff or my rnd macros for like 2 years
now. It's sad that I sunk so much time towards them and it ultimately
didn't work out, but typst is just so much better in every conceivable
way. And roff never had a bright future ahead of it, it's too obscure.