[river] More migration progress
- run bar at startup
- add swayidle & waylock

I'm keeping physlock on the system (in api script), but I moved away
from it in my scripts, because I discovered that the issues I've been
getting on my NVIDIA machine (unable to resume after suspend) are solved
when I use waylock and am able to switch TTYs to wake up the monitor. I
don't know why it happens, this is a dirty workaround. But it makes it
possible to wake the system without shutting it down with the power
button, which is, well, important for me.