[river] Migrate from pactl to amixer for volume
amixer is ALSA, more generic. pactl is an unnecessary PulseAudio
dependency. I did, however, notice that amixer behaves less consistently
and when having the keybindings do changes every 2%, sometimes it goes
up or down by 1% or 3%, which is annoying. To pretend this doesn't
happen, I changed the sensitivity of the keybindings to change every 1%.